Handle incoming batches the right way
The way Kudu scans were handled was involving doing one large scan on the scanner token at once. In fact, the way
most data stores work is that they define maximum number of records, that are later merged together. This was fixed
to follow the standards and now seems to be working fine.